+++ src/models/st7735s.rs
@@ -28,6 +28,8 @@ impl Model for ST7735s {
         DELAY: DelayNs,
         DI: Interface,
     {
+        assert_interface_kind!(Serial4Line | Parallel8Bit | Parallel16Bit);
+
         let madctl = SetAddressMode::from(options);
 
         delay.delay_us(200_000);